+++ src/Oro/Bundle/ApiBundle/Tests/Functional/JsonApiDocContainsConstraint.php
@@ -3,7 +3,6 @@
 namespace Oro\Bundle\ApiBundle\Tests\Functional;
 
 use Oro\Bundle\ApiBundle\Request\JsonApi\JsonApiDocumentBuilder as JsonApiDoc;
-use Oro\Component\PhpUtils\ArrayUtil;
 use Oro\Component\Testing\Assert\ArrayContainsConstraint;
 
 /**
@@ -46,8 +45,8 @@ class JsonApiDocContainsConstraint extends ArrayContainsConstraint
                     $expectedItems = $this->getDataItems($expectedData);
                     $actualItems = $this->getDataItems($other[JsonApiDoc::DATA]);
                     if (!$this->strictPrimaryData) {
-                        ArrayUtil::sortBy($expectedItems, false, JsonApiDoc::ID, SORT_STRING);
-                        ArrayUtil::sortBy($actualItems, false, JsonApiDoc::ID, SORT_STRING);
+                        $this->sortDataForNotStrictComparison($expectedItems);
+                        $this->sortDataForNotStrictComparison($actualItems);
                     }
                     try {
                         \PHPUnit\Framework\Assert::assertSame(
@@ -134,5 +133,21 @@ class JsonApiDocContainsConstraint extends ArrayContainsConstraint
         }
 
         return $result;
+    }
+
+    private function sortDataForNotStrictComparison(array &$data): void
+    {
+        $sortingData = [];
+        foreach ([JsonApiDoc::TYPE, JsonApiDoc::ID] as $fieldName) {
+            $sortingItems = array_column($data, $fieldName);
+            if ($sortingItems) {
+                $sortingData[] = array_column($data, $fieldName);
+                $sortingData[] = SORT_ASC;
+            }
+        }
+        if ($sortingData) {
+            $sortingData[] = &$data;
+            array_multisort(...$sortingData);
+        }
     }
 }