+++ src/Oro/Bundle/ApiBundle/Tests/Functional/RestJsonApiCustomization/CustomCompositeIdentifierTest.php
@@ -270,7 +270,7 @@ class CustomCompositeIdentifierTest extends RestJsonApiTestCase
 
         $response = $this->cget(
             ['entity' => $entityType],
-            ['filter[id]!' => $this->getEntityId('item 3', 3)]
+            ['filter[id][neq]' => $this->getEntityId('item 3', 3)]
         );
 
         $this->assertResponseContains(
@@ -348,7 +348,7 @@ class CustomCompositeIdentifierTest extends RestJsonApiTestCase
 
         $response = $this->cget(
             ['entity' => $entityType],
-            ['filter[id]!' => $this->getEntityId('item 1', 1) . ',' . $this->getEntityId('item 3', 3)]
+            ['filter[id][neq]' => $this->getEntityId('item 1', 1) . ',' . $this->getEntityId('item 3', 3)]
         );
 
         $this->assertResponseContains(